+++ src/utility.h
@@ -3,8 +3,8 @@
 #include <stdlib.h>
 #include <math.h>
 
-int   randm    (int);
-int   nmod     (int, int);
-float perlin2d (int, double, double, double);
-float dist2d   (float, float, float, float);
-float dist3d   (float, float, float, float, float, float);
+int    randm    (int);
+int    nmod     (int, int);
+double perlin2d (int, double, double, double);
+double dist2d   (double, double, double, double);
+double dist3d   (double, double, double, double, double, double);
